Split out `BeatmapCarousel` classes and drop `V2` suffix
